+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2 [unreleased]
 
 - Added Mailchimp service
+- Fixed Mandrill service
 - Added `uniq` to `subscribed` scope
 
 ## 0.0.1